increase length of laser pulses

+++ Marlin/planner.cpp
@@ -525,7 +525,7 @@ float junction_deviation = 0.1;
 #ifdef ENABLE_AUTO_BED_LEVELING
 void plan_buffer_line(float x, float y, float z, const float &e, float feed_rate, const uint8_t &extruder)
 #else
-void plan_buffer_line(const float &x, const float &y, const float &z, const float &e, float feed_rate, const uint8_t &extruder, bool laser, float laser_power, float laser_ppm, int8_t laser_pulse)
+void plan_buffer_line(const float &x, const float &y, const float &z, const float &e, float feed_rate, const uint8_t &extruder, bool laser, float laser_power, float laser_ppm, long laser_pulse)
 #endif  //ENABLE_AUTO_BED_LEVELING
 {
   // Calculate the buffer head after we push this byte
